engine/requirements.txt
# Core FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Azure AI services
azure-ai-documentintelligence==1.0.0b1
azure-core==1.29.5
azure-identity==1.15.0

# LangChain and OpenAI
langchain==0.0.352
langchain-community==0.0.13
langchain-openai==0.0.2
openai==1.3.8

# Document processing
python-docx==1.1.0
pypdf2==3.0.1
pdfplumber==0.10.3
pytesseract==0.3.10
pillow==10.1.0
openpyxl==3.1.2

# PDF generation
fpdf2==2.7.6
reportlab==4.0.7

# Environment and utilities
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4

# Flask (for the alternative app)
flask==3.0.0
werkzeug==3.0.1

# Additional utilities
numpy==1.25.2
opencv-python==4.8.1.78
